:recycle: refactor: rename graphql-codegen package
