+++ apps/web/next.config.js
@@ -1,5 +1,5 @@
 const withTM = require("next-transpile-modules")([
-  "@misterdonut/graphql-codegen",
+  "@misterdonut/graphql",
   "@misterdonut/ui",
 ]);
