+++ packages/ui/package.json
@@ -18,7 +18,7 @@
     "@emotion/cache": "11.10.3",
     "@emotion/react": "11.10.4",
     "@emotion/styled": "11.10.4",
-    "@misterdonut/graphql-codegen": "*",
+    "@misterdonut/graphql": "*",
     "@mui/icons-material": "5.10.6",
     "@mui/material": "5.10.6",
     "@mui/x-data-grid": "5.17.4",
